# openapi.yml
openapi: 3.0.3
info:
  title: SIGC - Sistema de Gestión de Citas Médicas
  description: API completa para gestión de especialidades, doctores, citas y usuarios
  version: 1.0.0
  contact:
    name: Soporte SIGC
    email: soporte@sigc.com
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo
  - url: https://api.sigc.com
    description: Servidor de producción
...
paths:
...
  /citas:
    get:
      tags:
        - Citas
      summary: Listar citas
      description: Retorna citas del usuario autenticado o todas las citas (si es admin)
      security:
        - BearerAuth: []
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [PENDIENTE, CONFIRMADA, CANCELADA, COMPLETADA]
          description: Filtrar por estado
        - name: fecha
          in: query
          schema:
            type: string
            format: date
          description: Filtrar por fecha (YYYY-MM-DD)
      responses:
        '200':
          description: Lista de citas obtenida
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CitaResponse'
              example:
                - idCita: 1
                  paciente: Juan Pérez
                  doctor: Dr. Carlos López
                  fecha: "2025-12-15"
                  hora: "10:30"
                  especialidad: Cardiología
                  estado: CONFIRMADA
        '401':
          description: No autorizado

    post:
      tags:
        - Citas
      summary: Crear nueva cita
      description: Reserva una nueva cita médica
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearCitaRequest'
            example:
              idDoctor: 1
              fecha: "2025-12-15"
              hora: "10:30"
              motivo: Revisión periódica
      responses:
        '201':
          description: Cita creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitaResponse'
        '400':
          description: Datos inválidos
        '401':
          description: No autorizado
        '409':
          description: Horario no disponible
  /citas/{id}:
    get:
      tags:
        - Citas
      summary: Obtener cita por ID
      description: Retorna los detalles de una cita específica
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Cita encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitaResponse'
        '404':
          description: Cita no encontrada
        '401':
          description: No autorizado

    put:
      tags:
        - Citas
      summary: Actualizar cita
      description: Modifica los detalles de una cita
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarCitaRequest'
      responses:
        '200':
          description: Cita actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitaResponse'
        '404':
          description: Cita no encontrada
        '401':
          description: No autorizado

    delete:
      tags:
        - Citas
      summary: Cancelar cita
      description: Cancela una cita existente
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Cita cancelada exitosamente
        '404':
          description: Cita no encontrada
        '401':
          description: No autorizado
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido al hacer login
  schemas:
...
    CitaResponse:
      type: object
      properties:
        idCita:
          type: integer
          example: 1
        paciente:
          type: string
          example: Juan Pérez
        doctor:
          type: string
          example: Dr. Carlos López
        fecha:
          type: string
          format: date
          example: "2025-12-15"
        hora:
          type: string
          example: "10:30"
        especialidad:
          type: string
          example: Cardiología
        motivo:
          type: string
          example: Revisión periódica
        estado:
          type: string
          enum: [PENDIENTE, CONFIRMADA, CANCELADA, COMPLETADA]
          example: CONFIRMADA
        fechaCreacion:
          type: string
          format: date-time
    CrearCitaRequest:
      type: object
      required:
        - idDoctor
        - fecha
        - hora
      properties:
        idDoctor:
          type: integer
          example: 1
        fecha:
          type: string
          format: date
          description: Fecha en formato YYYY-MM-DD
          example: "2025-12-15"
        hora:
          type: string
          description: Hora en formato HH:MM
          example: "10:30"
        motivo:
          type: string
          example: Revisión periódica
    ActualizarCitaRequest:
      type: object
      properties:
        fecha:
          type: string
          format: date
        hora:
          type: string
        motivo:
          type: string
        estado:
          type: string
          enum: [PENDIENTE, CONFIRMADA, CANCELADA, COMPLETADA]
...